Rewrite `pt_renaissance` with instrumenting agent
The Renaissance benchmark suite's launcher runs each benchmark with its
own class loader to prevent the interference between the various
benchmarks' dependencies.

Instrumentation sites within the benchmarks need to be able to load our
code. Just adding our code to the system (a.k.a application) class path
did not make it available within the benchmarks. My current solution to
this problem was to append to the boot (a.k.a primordial) class path.

The rules in `@dwtj_rules_java` aren't yet flexible enough to do this,
so for this prototype, I created a `sh_test` rule that invokes `java`.
See `@pt_renaissance//pt/agent:instrument_renaissance_benchmark_test`.