Use `$(location)` in the `pt_renaissance` test
This simplifies this test script: file paths (relative to the test
script's execution directory) are now given to the script by Bazel
rather than hard-coded in the script.

The `@dwtj_rules_java` version was bumped to support this. This new
version was needed to use `$(location @dwtj_local_openjdk//:java)`.