Move asm & renaissance deps from pt to root wkspc
It looks like I'm going to be using both of these dependencies in
various prototypes and the main project, so I've moved them out of the
`@pt_instrument_renaissance` architecture prototype workspace and into
the root `@phasicj` workspace. This prototype as refactored to use these
dependencies from `@phasicj//third_party/...`.