Add instr cli stub; move instr SVM code to subpkg
The PhasicJ Agent's instrumentation code is currently integrated into the
agent using `native-image`, SVM, and GraalVM isolates. This integration
is currently a bit of a mess. In particular, the agent appears to be
consistently segfaulting while instrumenting a particular Renaissance
benchmark, `apache-spark`. However, I'm having a hard time tracking down
the cause. (E.g., I don't know what program point is causing the
segfault, and I'm not sure which is class is triggering the segfault.)

In this commit, I've added a stub for a CLI application which wraps the
instrumentation logic. I hope that such a tool may help me diagnose
the problem.

I've also moved instr's SVM-specific wrapping code to a subpackage. This
seems appropriate, since there are now two ways to wrap/drive the
`instr` code: CLI and SVM.