Draft start of `pt_rust_grpc_ipc` prototype
This prototype project includes the beginnings of three "Hello, world!"
gRPC servers implemented in Rust. They differ in the details of how gRPC
service definitions are compiled to Rust code:

1. `Cargo/` uses Cargo and `tonic`.
2. `Bazel/pt/grpc/bazel_style` uses Bazel and `@rules_rust`.
3. `Bazel/pt/grpc/cargo_style` uses Bazel, a Cargo-style build script,
and `tonic`.

A notable difference discovered by this exploration is that the code
generated by `tonic` is `async`, whereas the code generated by
`@rules_rust` is not.

A problem with option 3 is that `protoc` and the protobuf native include
directory (i.e., protobuf headers) are not automatically discovered by
tonic while Bazel runs the Cargo-style build script. This was overcome
by adding some environment variables, but this solution is brittle.