Use pre-compiled Renaissance JAR, not from-source
For now, I'm abandoning my efforts to get the Renaissance from-source
code working elegantly. This work is still located in
`//third_party/renaissance/from_source:jar`, but it is commented out
Instead, I'm using a JAR which I pre-compiled (using this target
and manually uploaded this to GitHub.

Note that I'm only using a version built from source rather than the
official Renaissance Benchmarks v0.12.0 release binaries to avoid a bug
that I am seeing when these are run in my Linux CI container. (This bug
may be related to running the benchmarks as root.)

Here's the problem with the current version of `from_source:jar`: the
JAR is begin re-compiling unnecessarily. (It takes while to finish. On
my dev laptop it takes 140s.)

Here's my current understanding of what I think is wrong. Because it is
`local=True` modifications are being made in-source-tree. Then these
modifications are seen as changes to the inputs and thus signal that the
`genrule` should be re-run. I have tried fixing by ignoring these
changes using an exclude clause in `//bazel/external:BUILD.renaissance`,
but there is still something wrong. I may need to investigate exactly
what is changing more closely.

Note that I would like to remove `local=True` from this `genrule`, but
that causes other problems.