Make daemon record RT events in a SQLite DB file
Added simple SQLite support to the daemon via the `rusqlite` Cargo
crate.

Also added SQLite as a dependency within the environments of the CI
servers.

This is only a prototype to get started with this SQLite integration.
The way that SQLite has been integrated is very rudimentary and it has
very poor concurrency, as I pointed out in this comment:

    WARNING! This design has very bad concurrency. Different
    `record_events` gRPC calls will all contend for a single SQLite
    `Connection` object and they will block their threads from the Tokio
    runtime's thread pool while doing so. This is exactly what one is
    not supposed to do when using async/await and a thread-pool-based
    Tokio runtime. I've only implemented it this way temporarily to
    begin prototyping our use of SQLite via `rusqlite`.