Move rtevents gRPC generated code into one crate
Previously, I was using

```
tonic::include_proto!("phasicj.services.rtevents");
```

to include this generated code into a crate if it was needed. Now, this
generated code is only included in `//phasicj/services/rtevents/rust`
and other crates which want to use the generated code code depend upon
this crate in the normal way.